+++ source/Modules/AzureDevOpsDsc.Common/Resources/Functions/Public/AzDoOrganizationSettings/Get-AzDoOrganizationSettings.ps1
@@ -9,6 +9,12 @@ Function Get-AzDoOrganizationSettings
         [Parameter()][bool]$EnableOAuthAuthentication,
         [Parameter()][bool]$EnableSSHAuthentication,
         [Parameter()][bool]$DisallowAadGuestUserPolicy,
+        [Parameter()][bool]$EnableIPConditionalAccessPolicyValidation,
+        [Parameter()][bool]$LogAuditEvents,
+        [Parameter()][bool]$AllowTeamAdminsToInviteUsers,
+        [Parameter()][bool]$EnableRequestAccess,
+        [Parameter()][string]$RequestAccessUrl,
+        [Parameter()][bool]$EnableArtifactsFeedUpstreamProtection,
         [Parameter()][HashTable]$LookupResult,
         [Parameter()][Ensure]$Ensure,
         [Parameter()][System.Management.Automation.SwitchParameter]$Force
@@ -18,8 +24,9 @@ Function Get-AzDoOrganizationSettings
 
     $result = @{ Ensure = [Ensure]::Present; propertiesChanged = @(); status = $null }
 
+    $apiUri = 'https://dev.azure.com/{0}/' -f (Get-AzDoOrganizationName)
     $params = @{
-        ApiUri = 'https://dev.azure.com/{0}/' -f (Get-AzDoOrganizationName)
+        ApiUri = $apiUri
     }
 
     try
@@ -47,6 +54,34 @@ Function Get-AzDoOrganizationSettings
         if ($PSBoundParameters.ContainsKey('EnableOAuthAuthentication')  -and $liveEnableOAuth                -ne $EnableOAuthAuthentication)  { $changed += 'EnableOAuthAuthentication' }
         if ($PSBoundParameters.ContainsKey('EnableSSHAuthentication')    -and $liveEnableSSH                  -ne $EnableSSHAuthentication)    { $changed += 'EnableSSHAuthentication' }
         if ($PSBoundParameters.ContainsKey('DisallowAadGuestUserPolicy') -and $liveDisallowAadGuestUserPolicy -ne $DisallowAadGuestUserPolicy) { $changed += 'DisallowAadGuestUserPolicy' }
+
+        # Organization policies (separate API: _apis/OrganizationPolicy/Policies/{policyName})
+        $policyMap = Get-DevOpsOrganizationPolicyMap
+        $livePolicyValues = @{}
+
+        foreach ($entry in $policyMap)
+        {
+            $policy = Get-DevOpsOrganizationPolicy -ApiUri $apiUri -PolicyName $entry.PolicyName
+            $liveValue = [System.Boolean]$policy.value
+            $livePolicyValues[$entry.PropertyName] = $liveValue
+            $result.($entry.PropertyName) = $liveValue
+
+            if ($PSBoundParameters.ContainsKey($entry.PropertyName) -and $liveValue -ne (Get-Variable -Name $entry.PropertyName -ValueOnly))
+            {
+                $changed += $entry.PropertyName
+            }
+
+            if ($entry.PropertyName -eq 'EnableRequestAccess')
+            {
+                $liveRequestAccessUrl = [string]$policy.url
+                $result.RequestAccessUrl = $liveRequestAccessUrl
+
+                if ($PSBoundParameters.ContainsKey('RequestAccessUrl') -and $PSBoundParameters.ContainsKey('EnableRequestAccess') -and $EnableRequestAccess -and $liveRequestAccessUrl -ne $RequestAccessUrl)
+                {
+                    $changed += 'RequestAccessUrl'
+                }
+            }
+        }
 
         $result.propertiesChanged = $changed
         $result.status = if ($changed.Count -eq 0) { [DSCGetSummaryState]::Unchanged } else { [DSCGetSummaryState]::Changed }